// src/openapi/models/event-request-log-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { AppClientSummaryDto } from './app-client-summary-dto';

/**
 * 
 * @export
 * @interface EventRequestLogDto
 */
export interface EventRequestLogDto {
    /**
     * 
     * @type {AppClientSummaryDto}
     * @memberof EventRequestLogDto
     */
    appClientUser?: AppClientSummaryDto;
    /**
     * 
     * @type {string}
     * @memberof EventRequestLogDto
     */
    eventType?: EventRequestLogDtoEventTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof EventRequestLogDto
     */
    eventCount?: number;
    /**
     * 
     * @type {boolean}
     * @memberof EventRequestLogDto
     */
    wasSuccessful?: boolean;
    /**
     * 
     * @type {string}
     * @memberof EventRequestLogDto
     */
    reason?: string;
    /**
     * 
     * @type {string}
     * @memberof EventRequestLogDto
     */
    lastAttempted?: string;
}

/**
    * @export
    * @enum {string}
    */
export enum EventRequestLogDtoEventTypeEnum {
    PersonChange = 'PERSON_CHANGE',
    PersonDelete = 'PERSON_DELETE',
    OrganizationChange = 'ORGANIZATION_CHANGE',
    OrganizationDelete = 'ORGANIZATION_DELETE',
    PersonOrgAdd = 'PERSON_ORG_ADD',
    PersonOrgRemove = 'PERSON_ORG_REMOVE',
    SubOrgAdd = 'SUB_ORG_ADD',
    SubOrgRemove = 'SUB_ORG_REMOVE'
}